add tests, fix events names & json.del only on real del

+++ src/lib.rs
@@ -79,18 +79,22 @@ fn json_del(ctx: &Context, args: Vec<String>) -> RedisResult {
     let redis_key = ctx.open_key_writable(&key);
     let deleted = match redis_key.get_value::<RedisJSON>(&REDIS_JSON_TYPE)? {
         Some(doc) => {
-            let res = if path == "$" {
+            if path == "$" {
                 redis_key.delete()?;
                 1
             } else {
                 doc.delete_path(&path)?
-            };
-            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_del", key.as_str());
-            ctx.replicate_verbatim();
-            res
+            }
         }
         None => 0,
     };
+
+    // no need to notify or replicate if nothing was deleted
+    if deleted > 0 {
+        ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.del", key.as_str());
+        ctx.replicate_verbatim();
+    }
+
     Ok(deleted.into())
 }
 
@@ -124,7 +128,7 @@ fn json_set(ctx: &Context, args: Vec<String>) -> RedisResult {
     match (current, set_option) {
         (Some(ref mut doc), ref op) => {
             if doc.set_value(&value, &path, op, format)? {
-                ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_set", key.as_str());
+                ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.set", key.as_str());
                 ctx.replicate_verbatim();
                 REDIS_OK
             } else {
@@ -136,7 +140,7 @@ fn json_set(ctx: &Context, args: Vec<String>) -> RedisResult {
             let doc = RedisJSON::from_str(&value, format)?;
             if path == "$" {
                 redis_key.set_value(&REDIS_JSON_TYPE, doc)?;
-                ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_set", key.as_str());
+                ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.set", key.as_str());
                 ctx.replicate_verbatim();
                 REDIS_OK
             } else {
@@ -264,14 +268,14 @@ fn json_type(ctx: &Context, args: Vec<String>) -> RedisResult {
 /// JSON.NUMINCRBY <key> <path> <number>
 ///
 fn json_num_incrby(ctx: &Context, args: Vec<String>) -> RedisResult {
-    json_num_op(ctx, "json_incrby", args, |i1, i2| i1 + i2, |f1, f2| f1 + f2)
+    json_num_op(ctx, "json.numincrby", args, |i1, i2| i1 + i2, |f1, f2| f1 + f2)
 }
 
 ///
 /// JSON.NUMMULTBY <key> <path> <number>
 ///
 fn json_num_multby(ctx: &Context, args: Vec<String>) -> RedisResult {
-    json_num_op(ctx, "json_multby", args, |i1, i2| i1 * i2, |f1, f2| f1 * f2)
+    json_num_op(ctx, "json.nummultby", args, |i1, i2| i1 * i2, |f1, f2| f1 * f2)
 }
 
 ///
@@ -280,7 +284,7 @@ fn json_num_multby(ctx: &Context, args: Vec<String>) -> RedisResult {
 fn json_num_powby(ctx: &Context, args: Vec<String>) -> RedisResult {
     json_num_op(
         ctx,
-        "json_numpowby",
+        "json.numpowby",
         args,
         |i1, i2| i1.pow(i2 as u32),
         |f1, f2| f1.powf(f2),
@@ -314,7 +318,7 @@ fn json_bool_toggle(ctx: &Context, args: Vec<String>) -> RedisResult {
             .map_err(|e| e.into())
         })
         .and_then(|v: RedisValue| {
-            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_toggle", key.as_str());
+            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.toggle", key.as_str());
             ctx.replicate_verbatim();
             Ok(v)
         })
@@ -435,7 +439,7 @@ fn json_str_append(ctx: &Context, args: Vec<String>) -> RedisResult {
             .map_err(|e| e.into())
         })
         .and_then(|v: RedisValue| {
-            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_strappend", key.as_str());
+            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.strappend", key.as_str());
             ctx.replicate_verbatim();
             Ok(v)
         })
@@ -487,7 +491,7 @@ fn json_arr_append(ctx: &Context, args: Vec<String>) -> RedisResult {
             .map_err(|e| e.into())
         })
         .and_then(|v: RedisValue| {
-            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_arrappend", key.as_str());
+            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.arrappend", key.as_str());
             ctx.replicate_verbatim();
             Ok(v)
         })
@@ -567,7 +571,7 @@ fn json_arr_insert(ctx: &Context, args: Vec<String>) -> RedisResult {
             .map_err(|e| e.into())
         })
         .and_then(|v: RedisValue| {
-            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_arrinsert", key.as_str());
+            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.arrinsert", key.as_str());
             ctx.replicate_verbatim();
             Ok(v)
         })
@@ -639,7 +643,7 @@ fn json_arr_pop(ctx: &Context, args: Vec<String>) -> RedisResult {
             .map_err(|e| e.into())
         })
         .and_then(|v| {
-            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_arrpop", key.as_str());
+            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.arrpop", key.as_str());
             ctx.replicate_verbatim();
             Ok(v)
         })?;
@@ -705,7 +709,7 @@ fn json_arr_trim(ctx: &Context, args: Vec<String>) -> RedisResult {
             .map_err(|e| e.into())
         })
         .and_then(|v: RedisValue| {
-            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json_arrtrim", key.as_str());
+            ctx.notify_keyspace_event(NotifyEvent::MODULE, "json.arrtrim", key.as_str());
             ctx.replicate_verbatim();
             Ok(v)
         })
